chore: comment invoking payment methods in cart
