+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/__tests__/applePayExpress.test.js
@@ -763,6 +763,7 @@ describe('onShippingMethodSelected function', () => {
         },
       }),
     };
+    global.fetch = jest.fn().mockResolvedValueOnce(mockCalculationResponse)
     selectShippingMethod = jest.fn().mockImplementation((data, resolveApplePay) => {
       return mockCalculationResponse
     });
