+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/cart.js
@@ -1,0 +1,7 @@
+/* eslint-disable global-require */
+const processInclude = require('base/util');
+
+$(document).ready(() => {
+  processInclude(require('base/cart/cart'));
+  processInclude(require('./cart/expressPayments'));
+});
