+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/cart/expressPayments.js
@@ -1,0 +1,14 @@
+// const { getPaymentMethods } = require('../commons');
+
+function init() {
+  $(document).ready(async () => {
+    $.spinner().start();
+    // const paymentMethods = await getPaymentMethods();
+    // const paymentMethodsResponse = await paymentMethods.json();
+    $.spinner().stop();
+  });
+}
+
+module.exports = {
+  init,
+};
